requirements.txt
folium==0.19.5
matplotlib==3.10.1
pandas==2.2.3
protobuf==4.25.3
pyproj==3.7.1
pytz==2024.1
requests==2.32.3
seaborn==0.13.2
geopandas==1.0.1
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-bigquery
pyarrow
pandas-gbq db-dtypes